+++ examples/simple.rs
@@ -59,6 +59,9 @@ fn main() {
     base_widget.on_mouse_up_inside(Box::new(|widget_id, button| {
         eprintln!("Mouse button released (inside same widget): {:?}", button);
     }));
+    base_widget.on_mouse_up_outside(Box::new(|widget_id, button| {
+        eprintln!("Mouse button release (outside widget): {:?}", button);
+    }));
     prod.widget_store.add_widget(Box::new(base_widget));
 
     let mut box_widget = BoxWidget::new();