+++ src/core/main.rs
@@ -158,15 +158,22 @@ impl Pushrod {
                         .insert(args.button);
 
                     self.widget_store.button_down(last_widget_id, args.button);
-                }
+                },
                 ButtonState::Release => {
-                    let mut button_set = button_map.entry(last_widget_id)
+                    let button_set = button_map
+                        .entry(last_widget_id)
                         .or_insert(HashSet::new());
 
                     if button_set.contains(&args.button) {
                         button_set.remove(&args.button);
                         self.widget_store.button_up_inside(last_widget_id, args.button);
                     } else {
+                        for (widget_id, button_set) in button_map.iter_mut() {
+                            if button_set.contains(&args.button) {
+                                self.widget_store.button_up_outside(*widget_id, args.button);
+                                button_set.remove(&args.button);
+                            }
+                        }
                         // Find the button that was set for button down in the button map
                         // search each set, and check for the args.button
                         // The hash_map that matches that ID should be called with a